requirements.txt
mcp
openai
python-dotenv
rich